net: eth-phy-uclass: add ability to bind an eth-dev to a eth-phy
5fe419ef2a ("net: Add eth phy generic driver for shared MDIO")
added the ability to bind a phy driver to an eth device for
shared MII bus access. This was implemented for the case
where an eth dev had a direct child phy node.

This patch adds the ability to bind a phy to an eth dev which
is not a direct child of that phy, such as the case of an
i2c or spi connected ethernet switch.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>