arm: dts: imx8mm-venice-gw7901: add rohm,reset-snvs-powered prop
The PMIC was designed to be used with NXP i.MX8 SoC and it supports
SNVS low power state which seems to be typical for NXP i.MX SoCs.
However, when SNVS is used we must not allow SW to control
enabling/disabling those regulators which are crucial for system
to boot as there is a HW limitation which causes SW controlled
regulators to be kept shut down after SNVS reset.

Add rohm,reset-snvs-powered prop to use the SNVS reset state
and mark the regulators which are critical for boot.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>