board: gateworks: venice: fix SDRAM fixup
Use fdt to determine DRAM size. When support for the IMX8M Nano was
introduced it changed to using the fdt to store the proper dram size
instead of stuffing it in M4_BOOTROM_BASE_ADDR which was a hack that
didn't work on IXM8M Nano. With that change the ft_board_setup()
function also needed an update.

Fixes 94997d622b ("board: gateworks: venice: add imx8mn-gw7902 support")

Signed-off-by: Tim Harvey <tharvey@gateworks.com>