Revert "arm: dts: imx8mp-venice-gw74xx: flip ethernet aliases to match Linux"
This reverts commit 0f234dfd2544e4ccf6e6844db20b0ab69830e4f3.

Linux currently does not provide determinate network interface naming
for platform network devices so fec/eqos devices will register in a
random order and flip between which one is eth0 and eth1.

To combat this one is supposed to use systemd/udev to rename the
interfaces to their liking in userspace.

To avoid confusing of interfaces changing names between kernel init and
userspace, and to remove the dependence on systemd/udev we have patched
the kernel to use the interface index from dt aliases like U-Boot does
so now we can define the gw74xx ethernet interfaces as desired:
  eqos: eth0 (this is the first RJ45 out of the 6 available)
  fec:  eth1 (this is the cpu uplink to the KSZ switch)

Signed-off-by: Tim Harvey <tharvey@gateworks.com>