usb: hub: leave power domain enabled to avoid hang on host controller remove
When a USB host controller is removed that is configured for HOST mode
the usb_hub driver will disable its power domains and the subsequent
call to ehci_shutdown from the host controller driver will hang.

Set the flag to leave power domain on for this device/driver to avoid
this.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>