imx8mp-venice-gw82xx-2x: add gpio hog for usb1 out front panel for SDP
The imx8mp-venice-gw82xx-2x has a MUX that routes USB1 to the front
panel Type-C connector when GPIO4_IO17 is low and to J29 when driven high.

The reason for this is to allow SDP support via the front-panel
connector when the board is powered up.

Add a gpio-hog to drive the gpio high after boot to route USB1 to J29
for normal board use.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>