usb: fix usb gadget registration
Avoid infinite recursion caused by removing existing usb host when
registering a gadget. It's not our responsibility to unbind existing
devices.

In the case of an imx8mm with ci_udc binding to usb_ether results in the
following:
 usb           0  [   ]   ehci_mx6              |   |   |-- usb@32e40000
 ethernet      1  [   ]   usb_ether             |   |   |   `-- usb@32e40000
 bootdev       2  [   ]   eth_bootdev           |   |   |       `-- usb@32e40000.bootdev
 usb           1  [   ]   ehci_mx6              |   |   |-- usb@32e50000

When the usb_ether device is used the driver probes and
usb_setup_ehci_gadget calls device_remove on usb@32e4000 which removes
its children causing a recursion issue.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>