net: mv88e6xxx: add support for MV88E6361
The MV88E6361 port register base offset is 0x0. Taking a cue from the
linux kernel driver, define base offset based on compatible family and
use this to determine the actual part.

If chip is not defined by compatible, fall back to the previous method
which tries to read the actual part id using two different sets of
registers.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>